+++ crates/rpc/src/handlers/checkpoint.rs
@@ -7,8 +7,11 @@ use warp::{
     reply::{Reply, with_status},
 };
 
-use super::{BeaconResponse, state::get_state_from_id};
-use crate::types::id::ID;
+use super::state::get_state_from_id;
+use crate::types::{
+    id::ID,
+    response::BeaconResponse,
+};
 
 #[derive(Debug, Serialize, Clone)]
 pub struct CheckpointData {
