+++ crates/rpc/src/handlers/mod.rs
@@ -8,13 +8,11 @@ pub mod state;
 pub mod validator;
 pub mod version;
 
-use alloy_primitives::B256;
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use serde_json::json;
 use warp::reply::{Json, json};
 
-const EXECUTION_OPTIMISTIC: bool = false;
-const FINALIZED: bool = false;
+
 
 /// A generic data struct that can be used to wrap any data type
 /// used for json rpc responses
@@ -34,41 +32,6 @@ impl<T: Serialize> Data<T> {
     }
 }
 
-/// A BeaconResponse data struct that can be used to wrap data type
-/// used for json rpc responses
-///
-/// # Example
-/// {
-///  "data": json!({
-///     "execution_optimistic" : bool,
-///     "finalized" : bool,
-///     "data" : T
-/// })
-/// }
-#[derive(Debug, Serialize)]
-pub struct BeaconResponse<T> {
-    pub execution_optimistic: bool,
-    pub finalized: bool,
-    pub data: T,
-}
 
-impl<T: Serialize> BeaconResponse<T> {
-    pub fn json(data: T) -> Json {
-        json(&json!(Self {
-            data,
-            execution_optimistic: EXECUTION_OPTIMISTIC,
-            finalized: FINALIZED
-        }))
-    }
-}
 
-#[derive(Serialize, Deserialize)]
-pub struct RootResponse {
-    pub root: B256,
-}
 
-impl RootResponse {
-    pub fn new(root: B256) -> Self {
-        Self { root }
-    }
-}
