+++ crates/rpc/src/types/response.rs
@@ -1,5 +1,10 @@
 use alloy_primitives::B256;
 use serde::{Deserialize, Serialize};
+use serde_json::json;
+use warp::reply::{Json, json};
+
+const EXECUTION_OPTIMISTIC: bool = false;
+const FINALIZED: bool = false;
 
 #[derive(Serialize, Deserialize)]
 pub struct RootResponse {
@@ -9,5 +14,33 @@ pub struct RootResponse {
 impl RootResponse {
     pub fn new(root: B256) -> Self {
         Self { root }
+    }
+}
+
+/// A BeaconResponse data struct that can be used to wrap data type
+/// used for json rpc responses
+///
+/// # Example
+/// {
+///  "data": json!({
+///     "execution_optimistic" : bool,
+///     "finalized" : bool,
+///     "data" : T
+/// })
+/// }
+#[derive(Debug, Serialize)]
+pub struct BeaconResponse<T> {
+    pub execution_optimistic: bool,
+    pub finalized: bool,
+    pub data: T,
+}
+
+impl<T: Serialize> BeaconResponse<T> {
+    pub fn json(data: T) -> Json {
+        json(&json!(Self {
+            data,
+            execution_optimistic: EXECUTION_OPTIMISTIC,
+            finalized: FINALIZED
+        }))
     }
 }
